ops_backend/ops_backend/asgi.py
"""
ASGI config for ops_backend project.

It exposes the ASGI callable as a module-level variable named ``application``.

For more information on this file, see
https://docs.djangoproject.com/en/4.1/howto/deployment/asgi/
"""

import os

from django.core.asgi import get_asgi_application

os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'ops_backend.settings.dev')

from channels.routing import ProtocolTypeRouter
from .channel import channel

application = ProtocolTypeRouter({
    "http": get_asgi_application(),  # Django's ASGI application to handle traditional HTTP requests
    "websocket": channel.application,  # WebSocket handler
})
