ops_backend/requirements.txt
coreapi==2.3.3
cffi==1.17.1
Deprecated==1.2.13
Django==4.1.3
django-auditlog==2.2.2
django-cors-headers==3.13.0
django-filter==22.1
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
eventlet==0.33.3
jsonpath==0.82
lxml==4.9.2
PyMySQL==1.0.2
pyparsing==3.0.9